Small memory optimization and move javadocs on the User.

+++ core/src/com/horse/pokemon/ObjectData/Players/PlayerActions.java
@@ -1,15 +1,15 @@
 package com.horse.pokemon.ObjectData.Players;
 
 enum PlayerActions {
-    IDLE(0f), WALKING(1f), RUNNING(2f), BIKING(3f), SWIMMING(1f), USING_HM_MOVE(0f), FISHING(0f), IN_BATTLE(0f);
+    IDLE(0), WALKING(1), RUNNING(2), BIKING(3), SWIMMING(1), USING_HM_MOVE(0), FISHING(0), IN_BATTLE(0);
     
-    float speed;
+    int speed;
     
-    PlayerActions(float speed) {
+    PlayerActions(int speed) {
         this.speed = speed;
     }
     
-    public float getSpeed() {
+    public int getSpeed() {
         return speed;
     }
 }
